requirements.txt
nump==5.5.5.5
numpy==2.2.4
pyqtgraph==0.13.7
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
pyzmq==26.4.0
scipy==1.15.2
shiboken6==6.9.0
